requirements.txt
aws-cdk-lib==2.130.0
constructs>=10.0.0
boto3==1.39.4